+++ src/main/java/step3/model/LottoService.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.model;
 
 public class LottoService {
