+++ src/main/java/step3/model/Rank.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.model;
 
 import java.util.Arrays;
