+++ src/test/java/step3/LottosTest.java
@@ -4,6 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import step3.model.Lotto;
+import step3.model.Lottos;
+import step3.model.Rank;
 
 import java.util.ArrayList;
 import java.util.Arrays;
